blog_project/requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.11.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
audioop-lts==0.2.2
audioread==3.0.1
autopep8==2.3.2
babel==2.16.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.6
cryptograpy==0.0.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
deepface==0.0.95
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.17.0
fire==0.7.1
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
fsspec==2025.5.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
graphviz==0.21
grpcio==1.74.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.31.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.5.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-black==0.4.0
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_code_formatter==3.0.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.11.3
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.1
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml_dtypes==0.5.3
msgpack==1.1.2
mtcnn==1.0.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
namex==0.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numba==0.62.1
numpy==2.2.6
openai==1.82.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
pipenv==2025.0.4
platformdirs==4.3.6
plotly==6.2.0
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==6.31.1
psutil==6.1.1
psycopg2==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.11.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.2
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.22.3
scikit-learn==1.6.1
scikit-multilearn==0.2.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
simplejwt==2.0.1
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.6
soxr==1.0.0
sqlparse==0.5.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenize_rt==6.2.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.29.2
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.3
